Re-organize and cleanup
