+++ destination/destination.go
@@ -17,7 +17,7 @@ package destination
 import (
 	"context"
 
-	pubsub "github.com/conduitio-labs/conduit-connector-salesforce/pubsub"
+	pubsub "github.com/conduitio-labs/conduit-connector-salesforce/internal/pubsub"
 	"github.com/conduitio/conduit-commons/opencdc"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	"github.com/go-errors/errors"
