+++ internal/proto/buf.gen.yaml
@@ -2,7 +2,7 @@ version: v1
 managed:
   enabled: true
   go_package_prefix:
-    default: "github.com/conduitio-labs/conduit-connector-salesforce/proto"
+    default: "github.com/conduitio-labs/conduit-connector-salesforce/internal/proto"
 plugins:
   - plugin: buf.build/protocolbuffers/go:v1.34.1
     out: .
