+++ internal/pubsub/mock_pub_sub_subscribe_client_test.go
@@ -5,7 +5,7 @@ package pubsub
 import (
 	context "context"
 
-	eventbusv1 "github.com/conduitio-labs/conduit-connector-salesforce/proto/eventbus/v1"
+	eventbusv1 "github.com/conduitio-labs/conduit-connector-salesforce/internal/proto/eventbus/v1"
 	metadata "google.golang.org/grpc/metadata"
 
 	mock "github.com/stretchr/testify/mock"
